refactor : [Actions] added, [Operation], [Game] modified
