+++ src/main/java/Actions.java
@@ -1,0 +1,23 @@
+import java.util.Map;
+
+public class Actions {
+    private Condition condition = new Condition();
+
+    private void drive(Car car, Map<Car, Integer> carInfos) {
+        int currentPosition = carInfos.get(car);
+        carInfos.put(car, currentPosition + 1);
+    }
+
+    private void stop(Car car, Map<Car, Integer> carInfos) {
+        int currentPosition = carInfos.get(car);
+        carInfos.put(car, currentPosition);
+    }
+
+    void action(Car car, Map<Car, Integer> carInfos) {
+        if (condition.isGood()) {
+            drive(car, carInfos);
+            return;
+        }
+        stop(car, carInfos);
+    }
+}
